+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/internal-links/main.js
@@ -76,7 +76,7 @@ define([], function() {
 
                 this.sandbox.emit(
                     this.options.navigateEvent,
-                    'content/contents/' + this.options.webspace + '/' + this.options.locale + '/edit:' + id + '/details'
+                    'content/contents/' + this.options.webspace + '/' + this.options.locale + '/edit:' + id + '/content'
                 );
 
                 return false;
@@ -144,11 +144,14 @@ define([], function() {
         getColumnNavigationUrl = function() {
             var url = '/admin/api/nodes',
                 urlParts = [
-                    'webspace=' + this.options.webspace,
                     'language=' + this.options.locale,
                     'fields=title,order',
                     'webspace-nodes=all'
                 ];
+            
+            if (!!this.options.webspace) {
+                urlParts.push('webspace=' + this.options.webspace);
+            }
 
             return url + '?' + urlParts.join('&');
         },